+++ src/Codenizer.HttpClient.Testable/TestableHttpClientFactory.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Net.Http;
 
 namespace Codenizer.HttpClient.Testable
@@ -8,21 +9,33 @@ namespace Codenizer.HttpClient.Testable
     /// </summary>
     public class TestableHttpClientFactory : IHttpClientFactory
     {
-        private readonly Dictionary<string, TestableMessageHandler> _clientConfigurations = new Dictionary<string, TestableMessageHandler>();
+        private readonly Dictionary<string, TestableMessageHandler> _namedHandlers = new();
+        private readonly Dictionary<string, Action<System.Net.Http.HttpClient>> _clientConfigurations = new();
 
         /// <inheritdoc cref="IHttpClientFactory.CreateClient"/>
         public System.Net.Http.HttpClient CreateClient(string name)
         {
-            if (!_clientConfigurations.ContainsKey(name))
+            System.Net.Http.HttpClient httpClient;
+
+            if (!_namedHandlers.ContainsKey(name))
             {
                 // No client was configured but IHttpClientFactory.CreateClient
                 // requires us to return a new HttpClient instance. So we will
                 // return one but it's going to be backed by a default
                 // TestableMessageHandler. 
-                return new System.Net.Http.HttpClient(new TestableMessageHandler());
+                httpClient = new System.Net.Http.HttpClient(new TestableMessageHandler());
+            }
+            else
+            {
+                httpClient = new System.Net.Http.HttpClient(_namedHandlers[name]);
             }
 
-            return new System.Net.Http.HttpClient(_clientConfigurations[name]);
+            if (_clientConfigurations.ContainsKey(name))
+            {
+                _clientConfigurations[name](httpClient);
+            }
+
+            return httpClient;
         }
 
         /// <summary>
@@ -32,7 +45,21 @@ namespace Codenizer.HttpClient.Testable
         /// <param name="messageHandler">The <see cref="TestableMessageHandler"/> instance that should be used for the client name</param>
         public void ConfigureClient(string name, TestableMessageHandler messageHandler)
         {
-            _clientConfigurations.Add(name, messageHandler);
+            _namedHandlers.Add(name, messageHandler);
+        }
+
+        /// <summary>
+        /// Configure a client name to return a <see cref="HttpClient"/> with a <see cref="TestableMessageHandler"/> and configure the <see cref="HttpClient"/> instance with defaults
+        /// </summary>
+        /// <param name="name">The name of the <see cref="HttpClient"/> as requested by <see cref="CreateClient"/></param>
+        /// <param name="configure">A lambda to configure the created <see cref="HttpClient"/></param>
+        /// <returns>The <see cref="TestableMessageHandler"/> instance that will be used for the client name</returns>
+        public TestableMessageHandler ConfigureClient(string name, Action<System.Net.Http.HttpClient> configure)
+        {
+            _clientConfigurations.Add(name, configure);
+            var handler = ConfigureClient(name);
+
+            return handler;
         }
 
         /// <summary>
@@ -43,7 +70,7 @@ namespace Codenizer.HttpClient.Testable
         public TestableMessageHandler ConfigureClient(string name)
         {
             var messageHandler = new TestableMessageHandler();
-            _clientConfigurations.Add(name, messageHandler);
+            _namedHandlers.Add(name, messageHandler);
             return messageHandler;
         }
     }